+++ example/lib/demos/simple_shape/simple_shapes.dart
@@ -1,3 +1,5 @@
+import 'package:graphx/graphx.dart';
+
 import '../../utils/utils.dart';
 
 import 'simple_shapes_scene.dart';
@@ -7,7 +9,11 @@ class SimpleShapesMain extends StatelessWidget {
   Widget build(BuildContext context) {
     return DemoSingleSceneWidget(
       title: 'simple shapes drawing',
-      sceneRoot: SimpleShapesScene(),
+      root: SimpleShapesScene(),
+
+      /// we are rendering a static scene here, no need for Ticker
+      /// capabilities.
+      config: SceneConfig.static,
     );
   }
 }