+++ example/lib/demos/svg_icons/svg_icons_scene.dart
@@ -1,12 +1,7 @@
 import 'package:graphx/graphx.dart';
 import 'test_icons.dart';
 
-
-class SvgIconsScene extends SceneRoot {
-  SvgIconsScene() {
-    config(autoUpdateAndRender: true);
-  }
-
+class SvgIconsScene extends Sprite {
   @override
   void addedToStage() {
     addChild(TestIcons());