+++ example/lib/main.dart
@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:graphx/graphx.dart';
 import 'demos/demos.dart';
 
 void main() {
   runApp(
     MaterialApp(
-//      home: SimpleShapesMain(),
-//      home: SimpleTweenMain(),
-//      home: DemoSvgIconsMain(),
-//      home: SimpleInteractionsMain(),
-      home: GraphicsClipperDemo(),
+      // home: SimpleShapesMain(),
+      // home: SimpleTweenMain(),
+      // home: DemoSvgIconsMain(),
+      home: SimpleInteractionsMain(),
+//       home: GraphicsClipperDemo(),
     ),
   );
 }