+++ lib/src/core/core.dart
@@ -1,3 +1,4 @@
 export 'graphx.dart';
+export 'scene_config.dart';
 export 'scene_controller.dart';
 export 'scene_painter.dart';