+++ lib/src/core/scene_controller.dart
@@ -2,86 +2,15 @@ import 'package:flutter/material.dart';
 
 import '../../graphx.dart';
 
-class SceneConfig {
-  /// If the GraphX SceneController will use keyboard events.
-  bool useKeyboard;
-
-  /// If this GraphX SceneController will use pointer (touch/mouse) events.
-  bool usePointer;
-
-  /// Will be overwritten to `true` if [autoUpdateAndRender] is set on any
-  /// [ScenePainter] layer. Can be initialized on [SceneBuilderWidget()] or
-  /// [ScenePainter::setupScene()].
-  /// Warning: [setupScene()] takes priority over Widget initialization.
-  bool useTicker;
-
-  bool painterIsComplex;
-
-  /// See [CustomPaint.]
-  bool painterWillChange;
-
-  /// Avoids the scene from being disposed by the Widget.
-  /// Meant to be used with `ScenePainter.useOwnCanvas=true`
-  /// where a Layer captures it's own drawing to be used as Picture or Image
-  /// by other SceneLayers.
-  /// Warning: Experimental
-  bool isPersistent = false;
-
-  SceneConfig({
-    this.useKeyboard = false,
-    this.usePointer = false,
-    this.useTicker = false,
-    this.isPersistent = false,
-    this.painterIsComplex,
-    this.painterWillChange,
-  });
-
-  /// Quick method to set the object properties.
-  void setTo({
-    bool useTicker,
-    bool useKeyboard,
-    bool usePointer,
-    bool sceneIsComplex,
-  }) {
-    if (useTicker != null) this.useTicker = useTicker;
-    if (useKeyboard != null) this.useKeyboard = useKeyboard;
-    if (usePointer != null) this.usePointer = usePointer;
-    if (sceneIsComplex != null) painterIsComplex = sceneIsComplex;
-  }
-
-  /// Copy the values from another [SceneConfig] instance.
-  void copyFrom(SceneConfig other) {
-    useKeyboard = other.useKeyboard;
-    usePointer = other.usePointer;
-    useTicker = other.useTicker;
-    painterIsComplex = other.painterIsComplex;
-    painterWillChange = other.painterWillChange;
-    isPersistent = other.isPersistent;
-  }
-
-  /// Utility method used by the [SceneBuilderWidget] to set the flag
-  /// `CustomPaint.willChange`
-  bool painterMightChange() {
-    if (useTicker || usePointer || useKeyboard) {
-      return true;
-    }
-    return painterWillChange ?? false;
-  }
-}
-
 /// Entry point of GraphX world.
 /// A [SceneController] manages (up to) 2 [SceneRoot]s: `back` and `front`
 /// which correlates to [CustomPainter.painter] and
 /// [CustomPainter.foregroundPainter]. It takes care of the initialization
 /// and holding the references of the Painters used by [SceneBuilderWidget].
 class SceneController {
-  /// Eventually, might be a global access point to the current rendering
-  /// Scene, it should be in sync with the repainting process, so is safe
-  /// to assume that it will point to the proper SceneController reference.
-  // static SceneController current;
+  ScenePainter backScene;
 
-  ScenePainter back;
-  ScenePainter front;
+  ScenePainter frontScene;
 
   /// Access the `ticker` (if any) created by this SceneController.
   GxTicker get ticker {
@@ -103,7 +32,9 @@ class SceneController {
   PointerManager get pointer => _pointer;
 
   KeyboardManager _keyboard;
+
   PointerManager _pointer;
+
   GxTicker _ticker;
 
   InputConverter $inputConverter;
@@ -111,18 +42,49 @@ class SceneController {
   SceneConfig get config => _config;
 
   final _config = SceneConfig();
+
   int id = -1;
+
   bool _isInited = false;
 
+  set config(SceneConfig sceneConfig) {
+    _config.autoUpdateRender = sceneConfig.autoUpdateRender ?? true;
+    _config.isPersistent = sceneConfig.isPersistent ?? false;
+    _config.painterWillChange = sceneConfig.painterWillChange ?? true;
+    _config.useKeyboard = sceneConfig.useKeyboard ?? false;
+    _config.usePointer = sceneConfig.usePointer ?? false;
+    _config.useTicker = sceneConfig.useTicker ?? false;
+  }
+
+  Juggler get juggler => Graphx.juggler;
+
+  /// constructor.
+  SceneController({
+    Sprite back,
+    Sprite front,
+    SceneConfig config,
+  }) {
+    assert(back != null || front != null);
+    if (back != null) {
+      backScene = ScenePainter(this, back);
+    }
+    if (front != null) {
+      frontScene = ScenePainter(this, front);
+    }
+    this.config = config ?? SceneConfig.defaultConfig;
+  }
+
   /// WARNING: Internal method
-  /// starts from the `SceneBuilderWidget`.
+  /// called internally from [SceneBuilderWidget].
   void $init() {
-    if (_isInited) return;
+    if (_isInited) {
+      return;
+    }
     setup();
     if (_hasTicker()) {
       _ticker = GxTicker();
       _ticker.onFrame.add(_onTick);
-      if (_hasAutoUpdate()) {
+      if (_anySceneAutoUpdate()) {
         _ticker.resume();
       }
     }
@@ -131,19 +93,16 @@ class SceneController {
   }
 
   void setup() {
-    back?.$setup();
-    front?.$setup();
+    backScene?.$setup();
+    frontScene?.$setup();
   }
-
-  Juggler get juggler => Graphx.juggler;
 
   /// [GxTicker] that runs the `enterFrame`.
   /// Is independent from the rendering pipeline.
   void _onTick(double elapsed) {
     Graphx.juggler.update(elapsed);
-    // _makeCurrent();
-    front?.tick(elapsed);
-    back?.tick(elapsed);
+    frontScene?.tick(elapsed);
+    backScene?.tick(elapsed);
   }
 
   void resumeTicker() {
@@ -151,25 +110,18 @@ class SceneController {
   }
 
   void dispose() {
-    if (_config.isPersistent) return;
-    front?.dispose();
-    back?.dispose();
+    if (_config.isPersistent) {
+      return;
+    }
+    frontScene?.dispose();
+    backScene?.dispose();
     _ticker?.dispose();
     _ticker = null;
   }
 
-  set config(SceneConfig sceneConfig) {
-    _config.setTo(
-      sceneIsComplex: sceneConfig.painterIsComplex,
-      useKeyboard: sceneConfig.useKeyboard,
-      usePointer: sceneConfig.usePointer,
-      useTicker: sceneConfig.useTicker,
-    );
-  }
+  CustomPainter buildBackPainter() => backScene?.buildPainter();
 
-  CustomPainter buildBackPainter() => back?.buildPainter();
-
-  CustomPainter buildFrontPainter() => front?.buildPainter();
+  CustomPainter buildFrontPainter() => frontScene?.buildPainter();
 
   void _initInput() {
     if (_config.useKeyboard) {
@@ -183,35 +135,11 @@ class SceneController {
     }
   }
 
-  SceneController._();
+  bool _sceneAutoUpdate(ScenePainter scene) =>
+      scene?.autoUpdateAndRender ?? false;
 
-  static SceneController withLayers({
-    Sprite back,
-    Sprite front,
-    SceneConfig backConfig,
-    SceneConfig frontConfig,
-  }) {
-    assert(back != null || front != null);
-    final controller = SceneController._();
-    if (back != null) {
-      controller.back = ScenePainter(controller, back);
-      if (backConfig != null) {
-        controller.back.core.config = backConfig;
-      }
-    }
-    if (front != null) {
-      controller.front = ScenePainter(controller, front);
-      if (frontConfig != null) {
-        controller.back.core.config = frontConfig;
-      }
-    }
-    return controller;
-  }
+  bool _anySceneAutoUpdate() =>
+      _sceneAutoUpdate(backScene) || _sceneAutoUpdate(frontScene);
 
-  bool _layerHasAutoUpdate(ScenePainter p) => p?.autoUpdateAndRender ?? false;
-
-  bool _hasAutoUpdate() =>
-      _layerHasAutoUpdate(back) || _layerHasAutoUpdate(front);
-
-  bool _hasTicker() => _hasAutoUpdate() || _config.useTicker;
+  bool _hasTicker() => _anySceneAutoUpdate() || _config.useTicker;
 }